Flyttet skjema og komponenter ut i egne filer
