+++ frontend/mr-admin-flate/src/api/features/feature-toggles.ts
@@ -2,13 +2,19 @@ import { headers } from "../headers";
 import { useQuery } from "@tanstack/react-query";
 
 export const ENABLE_ADMIN_FLATE = "mulighetsrommet.enable-admin-flate";
+export const ENABLE_OPPRETT_TILTAKSTYPE =
+  "mulighetsrommet.enable-opprett-tiltakstype";
 
-export const ALL_TOGGLES = [ENABLE_ADMIN_FLATE] as const;
+export const ALL_TOGGLES = [
+  ENABLE_ADMIN_FLATE,
+  ENABLE_OPPRETT_TILTAKSTYPE,
+] as const;
 
 export type Features = Record<typeof ALL_TOGGLES[number], boolean>;
 
 export const initialFeatures: Features = {
   "mulighetsrommet.enable-admin-flate": false,
+  "mulighetsrommet.enable-opprett-tiltakstype": false,
 };
 
 const toggles = ALL_TOGGLES.map((element) => "feature=" + element).join("&");
