+++ frontend/mr-admin-flate/src/mocks/api/features.ts
@@ -2,4 +2,5 @@ import { Features } from "../../api/features/feature-toggles";
 
 export const mockFeatures: Features = {
   "mulighetsrommet.enable-admin-flate": true,
+  "mulighetsrommet.enable-opprett-tiltakstype": true,
 };
