+++ frontend/mr-admin-flate/src/pages/tiltakstyper/TiltakstyperPage.tsx
@@ -1,16 +1,22 @@
 import { BodyLong, Heading } from "@navikt/ds-react";
 import { Link } from "react-router-dom";
+import { useFeatureToggles } from "../../api/features/feature-toggles";
 import { TiltakstyperOversikt } from "../../components/tiltakstyper/TiltakstyperOversikt";
 import styles from "../tiltaksgjennomforinger/Oversikt.module.scss";
 
 export function TiltakstyperPage() {
+  const { data: toggles } = useFeatureToggles();
+
   return (
     <>
       <Link to="/">Hjem</Link>
       <Heading className={styles.overskrift} size="large">
         Oversikt over tiltakstyper
       </Heading>
-      <Link to="opprett">Opprett tiltakstype</Link>
+      {toggles?.["mulighetsrommet.enable-opprett-tiltakstype"] ? (
+        <Link to="opprett">Opprett tiltakstype</Link>
+      ) : null}
+
       <BodyLong className={styles.body} size="small">
         Her finner du dine aktive tiltakstyper.
       </BodyLong>
