+++ frontend/mr-admin-flate/src/pages/tiltakstyper/opprett-tiltakstyper/OpprettTiltakstypePage.tsx
@@ -1,161 +1,20 @@
-import {
-  BodyLong,
-  Button,
-  Checkbox,
-  Heading,
-  Select,
-  TextField,
-  UNSAFE_DatePicker,
-  UNSAFE_useRangeDatepicker,
-} from "@navikt/ds-react";
-import { FieldHookConfig, Form, Formik, useField } from "formik";
+import { BodyLong, Button, Heading } from "@navikt/ds-react";
+import { Form, Formik } from "formik";
 import { Link } from "react-router-dom";
-import { z } from "zod";
+
 import { toFormikValidationSchema } from "zod-formik-adapter";
-import styles from "../tiltaksgjennomforinger/Oversikt.module.scss";
+import styles from "../../tiltaksgjennomforinger/Oversikt.module.scss";
+import {
+  CheckboxFelt,
+  Datovelger,
+  SelectFelt,
+  Tekstfelt,
+} from "./OpprettTiltakstypeComponents";
 import formStyles from "./OpprettTiltakstypePage.module.scss";
-
-// TODO Skal vi ha auto-lagring av skjema?
-
-// TODO Dra skjema-validering ut i egen fil?
-const BooleanDefaultFalse = z.boolean().default(false);
-const Schema = z.object({
-  tiltakstypenavn: z.string({ required_error: "Tiltakstypen må ha et navn" }),
-  fraDato: z.date({
-    required_error: "Du må sette en fra-dato for tiltakstypen",
-  }),
-  tilDato: z.date({
-    required_error: "Du må sette en til-dato for tiltakstypen",
-  }),
-  tiltaksgruppekode: z.string({
-    required_error: "Tiltaksgruppekode må være satt",
-  }),
-  tiltakskode: z.string({ required_error: "Tiltakskode må være satt" }),
-  rettTilTiltakspenger: BooleanDefaultFalse,
-  administrasjonskode: z.string({
-    required_error: "Du må sette en administrasjonskode",
-  }),
-  kopiAvTilsagnsbrev: BooleanDefaultFalse,
-  arkivkode: z.string({ required_error: "Du må sette en arkivkode" }),
-  harAnskaffelse: BooleanDefaultFalse,
-  rammeavtale: z.string({ required_error: "Du må velge en rammeavtale" }),
-  opplaringsgruppe: z.string({
-    required_error: "Du må velge en opplæringsgruppe",
-  }),
-  handlingsplan: z.string({ required_error: "Du må velge en handlingsplan" }),
-  harObligatoriskSluttdato: BooleanDefaultFalse,
-  varighet: z.string({ required_error: "Du må sette en varighet" }),
-  harStatusSluttdato: BooleanDefaultFalse,
-  harStatusMeldeplikt: BooleanDefaultFalse,
-  harStatusVedtak: BooleanDefaultFalse,
-  harStatusIAAvtale: BooleanDefaultFalse,
-  harStatusTilleggstonad: BooleanDefaultFalse,
-  harStatusUtdanning: BooleanDefaultFalse,
-  harAutomatiskTilsagnsbrev: BooleanDefaultFalse,
-  harStatusBegrunnelseInnsok: BooleanDefaultFalse,
-  harStatusHenvisningsbrev: BooleanDefaultFalse,
-  harStatusKopibrev: BooleanDefaultFalse,
-});
-
-type OptionalSchemaValues = Partial<z.infer<typeof Schema>>;
-type SchemaValues = z.infer<typeof Schema>;
-
-// TODO Dra komponenter ut i egen fil?
-function Tekstfelt({
-  label,
-  name,
-  hjelpetekst,
-  ...props
-}: {
-  name: keyof SchemaValues;
-  label: string;
-  hjelpetekst?: string;
-} & FieldHookConfig<any>) {
-  const [field, meta] = useField({ name, ...props });
-  return (
-    <TextField
-      description={hjelpetekst}
-      size="small"
-      label={label}
-      {...field}
-      error={meta.touched && meta.error}
-    />
-  );
-}
-
-function SelectFelt({
-  label,
-  name,
-  defaultBlank = true,
-  defaultBlankName = "",
-  ...props
-}: {
-  name: keyof SchemaValues;
-  label: string;
-  defaultBlank?: boolean;
-  defaultBlankName?: string;
-} & FieldHookConfig<any>) {
-  const [field, meta] = useField({ name, ...props });
-  return (
-    <Select
-      size="small"
-      label={label}
-      {...field}
-      error={meta.touched && meta.error}
-    >
-      {defaultBlank ? <option value="">{defaultBlankName}</option> : null}
-      {props.children}
-    </Select>
-  );
-}
-
-function CheckboxFelt(
-  props: { name: keyof SchemaValues } & FieldHookConfig<any>
-) {
-  const [field] = useField({ ...props, type: "checkbox" });
-
-  return <Checkbox {...field}>{props.children}</Checkbox>;
-}
-
-// TODO Se på uthenting av dato på korrekt format
-function Datovelger() {
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const [fraDatoField, fraDatoMeta, fraDatoHelper] = useField("fraDato");
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const [tilDatoField, tilDatoMeta, tilDatoHelper] = useField("tilDato");
-  const { datepickerProps, toInputProps, fromInputProps } =
-    UNSAFE_useRangeDatepicker({
-      fromDate: new Date(),
-      onRangeChange: (val) => {
-        fraDatoHelper.setValue(val?.from);
-        tilDatoHelper.setValue(val?.to);
-      },
-    });
-  return (
-    <UNSAFE_DatePicker {...datepickerProps}>
-      <div style={{ display: "flex", gap: "5rem" }}>
-        <DatoFelt name="fraDato" label="Fra dato" {...fromInputProps} />
-        <DatoFelt name="tilDato" label="Til dato" {...toInputProps} />
-      </div>
-    </UNSAFE_DatePicker>
-  );
-}
-
-function DatoFelt({
-  name,
-  label,
-  ...rest
-}: { name: keyof SchemaValues; label: string } & FieldHookConfig<any> & any) {
-  const [_, meta] = useField({ name, ...rest });
-  return (
-    <UNSAFE_DatePicker.Input
-      {...rest}
-      label={label}
-      name={name}
-      error={meta.touched && meta.error}
-    />
-  );
-}
+import {
+  OpprettTiltakstypeSchema,
+  OptionalSchemaValues,
+} from "./OpprettTiltakstypeSchemaValidation";
 
 export function OpprettTiltakstype() {
   const initialValues: OptionalSchemaValues = {
@@ -193,10 +52,10 @@ export function OpprettTiltakstype() {
   const opplaringsgrupper = [
     { id: 1, navn: "Den beste opplæringsgruppen" },
     { id: 2, navn: "Den neste beste opplæringsgruppen" },
-  ]; // TODO Skal dette komme fra et api?
+  ]; // TODO Opplæringsgrupper må komme fra et API
   const handlingsplaner = [
     {
-      // TODO Skal dette komme fra et API?
+      // TODO Hva skal inn in en handlingsplan?
       id: 1,
       navn: "obligatorisk sluttdato",
     },
@@ -218,7 +77,7 @@ export function OpprettTiltakstype() {
       </BodyLong>
       <Formik<OptionalSchemaValues>
         initialValues={initialValues}
-        validationSchema={toFormikValidationSchema(Schema)}
+        validationSchema={toFormikValidationSchema(OpprettTiltakstypeSchema)}
         onSubmit={(values, actions) => {
           // TODO Må sende data til backend
           alert(JSON.stringify(values, null, 2));
